Move `testing.js` to root
Consolidate all entrypoints in the root of the project for consistency
and clarity.
