+++ .unimportedrc.json
@@ -1,5 +1,5 @@
 {
-  "entry": ["index.js", "src/testing.js"],
+  "entry": ["index.js", "testing.js"],
   "ignorePatterns": ["**/node_modules/**"],
   "ignoreUnimported": [],
   "ignoreUnused": [],
